+++ compiler/src/org.graalvm.compiler.truffle.options/src/org/graalvm/compiler/truffle/options/PolyglotCompilerOptions.java
@@ -26,6 +26,7 @@ package org.graalvm.compiler.truffle.options;
 
 import java.util.function.Function;
 
+import org.graalvm.collections.EconomicMap;
 import org.graalvm.options.OptionCategory;
 import org.graalvm.options.OptionDescriptors;
 import org.graalvm.options.OptionKey;
@@ -34,6 +35,9 @@ import org.graalvm.options.OptionType;
 import org.graalvm.polyglot.Engine;
 
 import com.oracle.truffle.api.Option;
+import java.util.Collections;
+import java.util.EnumSet;
+import java.util.Set;
 
 /**
  * Truffle compilation options that can be configured per {@link Engine engine} instance. These
@@ -60,6 +64,69 @@ public final class PolyglotCompilerOptions {
                             } catch (IllegalArgumentException e) {
                                 throw new IllegalArgumentException("Mode can be: 'default', 'latency' or 'throughput'.");
                             }
+                        }
+                    });
+
+    public enum PerformanceWarningKind {
+        VIRTUAL_RUNTIME_CALL("call", "Enables virtual call warnings"),
+        VIRTUAL_INSTANCEOF("instanceof", "Enables virtual instanceof warnings"),
+        VIRTUAL_STORE("store", "Enables virtual store warnings");
+
+        private static final EconomicMap<String, PerformanceWarningKind> kindByName;
+        static {
+            kindByName = EconomicMap.create();
+            for (PerformanceWarningKind kind : PerformanceWarningKind.values()) {
+                kindByName.put(kind.name, kind);
+            }
+        }
+
+        final String name;
+        final String help;
+
+        PerformanceWarningKind(String name, String help) {
+            this.name = name;
+            this.help = help;
+        }
+
+        public static PerformanceWarningKind forName(String name) {
+            PerformanceWarningKind kind = kindByName.get(name);
+            if (kind == null) {
+                throw new IllegalArgumentException("Unknown PerformanceWarningKind name " + name);
+            }
+            return kind;
+        }
+    }
+
+    static final OptionType<Set<PerformanceWarningKind>> PERFORMANCE_WARNING_TYPE = new OptionType<>("PerformanceWarningKind",
+                    new Function<String, Set<PerformanceWarningKind>>() {
+                        @Override
+                        public Set<PerformanceWarningKind> apply(String value) {
+                            if ("none".equals(value)) {
+                                return EnumSet.noneOf(PerformanceWarningKind.class);
+                            } else if ("all".equals(value)) {
+                                return EnumSet.allOf(PerformanceWarningKind.class);
+                            } else {
+                                Set<PerformanceWarningKind> result = EnumSet.noneOf(PerformanceWarningKind.class);
+                                for (String name : value.split(",")) {
+                                    try {
+                                        result.add(PerformanceWarningKind.forName(name));
+                                    } catch (IllegalArgumentException e) {
+                                        String message = String.format("The \"%s\" is not a valid performance warning kind. Valid values are%n", name);
+                                        for (PerformanceWarningKind kind : PerformanceWarningKind.values()) {
+                                            message = message + String.format("%s%s%s%n", kind.name, indent(kind.name.length()), kind.help);
+                                        }
+                                        message = message + String.format("all%sEnables all performance warnings%n", indent(3));
+                                        message = message + String.format("none%sDisables performance warnings%n", indent(4));
+                                        throw new IllegalArgumentException(message);
+                                    }
+                                }
+                                return result;
+                            }
+                        }
+
+                        private String indent(int nameLength) {
+                            int len = Math.max(1, 16 - nameLength);
+                            return new String(new char[len]).replace('\0', ' ');
                         }
                     });
 
@@ -132,7 +199,7 @@ public final class PolyglotCompilerOptions {
     public static final OptionKey<Boolean> CompilationExceptionsAreFatal = new OptionKey<>(false);
 
     @Option(help = "Treat performance warnings as fatal occurrences that will exit the applications", category = OptionCategory.INTERNAL)
-    public static final OptionKey<Boolean> PerformanceWarningsAreFatal = new OptionKey<>(false);
+    public static final OptionKey<Set<PerformanceWarningKind>> PerformanceWarningsAreFatal = new OptionKey<>(Collections.emptySet(), PERFORMANCE_WARNING_TYPE);
 
     // Tracing
 
@@ -261,7 +328,7 @@ public final class PolyglotCompilerOptions {
     public static final OptionKey<Boolean> InlineAcrossTruffleBoundary = new OptionKey<>(false);
 
     @Option(help = "Print potential performance problems", category = OptionCategory.INTERNAL)
-    public static final OptionKey<Boolean> TracePerformanceWarnings = new OptionKey<>(false);
+    public static final OptionKey<Set<PerformanceWarningKind>> TracePerformanceWarnings = new OptionKey<>(Collections.emptySet(), PERFORMANCE_WARNING_TYPE);
 
     @Option(help = "Prints a histogram of all expanded Java methods.", category = OptionCategory.INTERNAL)
     public static final OptionKey<Boolean> PrintExpansionHistogram = new OptionKey<>(false);